[preferences] Delay initialization
Calling `ContributionProvider.getContributions()` inside the constructor or
postConstruct of any component is just asking for troubles.

Issue is that until both the constructor and the postConstruct functions are
called, inversify won't see the component as "ready". In the case where a
contribution indirectly depends on a service relying on contributions, a
circular dependency loop will appear and will never be resolved in time.

By delaying the fetching of the contributions to after the component
construction/injection, it will be able to be injected back into its own
dependency.

This commit is required for the one after it to work.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>